+++ api/v1/routes.go
@@ -15,4 +15,10 @@ func SetupRoutes(router *gin.RouterGroup) {
 	{
 		pubGroup.POST("/action/", UserPublishHandler)
 	}
+
+	favoriteGroup := router.Group("/favorite")
+	{
+		favoriteGroup.POST("/action/", FavoriteAction)
+		favoriteGroup.GET("/list/", FavoriteList)
+	}
 }
