+++ database/models/video.go
@@ -5,14 +5,14 @@ import (
 )
 
 type Video struct {
-	VideoID       int64      `gorm:"primaryKey"`
-	AuthorUserID  int64      `gorm:"not null"`
-	PlayURL       string     `gorm:"size:256"`
-	CoverURL      string     `gorm:"size:256"`
-	Likes         int        `gorm:"default:0"`
-	Comments      int        `gorm:"default:0"`
-	Title         string     `gorm:"size:50"`
-	CreatedAt     time.Time  `gorm:"not null"`
-	UpdatedAt     *time.Time `gorm:"autoUpdateTime"`
-	DeletedAt     *time.Time `gorm:"index"`
+	VideoID      int64      `gorm:"column:video_id;primaryKey"`
+	AuthorUserID int64      `gorm:"column:author_user_id;not null"`
+	PlayURL      string     `gorm:"column:play_url;size:256"`
+	CoverURL     string     `gorm:"column:cover_url;size:256"`
+	Likes        int        `gorm:"column:likes;default:0"`
+	Comments     int        `gorm:"column:comments;default:0"`
+	Title        string     `gorm:"column:title;size:50"`
+	CreatedAt    time.Time  `gorm:"column:created_at;not null"`
+	UpdatedAt    *time.Time `gorm:"column:updated_at;autoUpdateTime"`
+	DeletedAt    *time.Time `gorm:"column:deleted_at;index"`
 }
